# Conflicts:
#	frontend/src/App.js
#	frontend/src/components/FormAdd.js

+++ backend/server/controllers/course_student/index.js
@@ -10,16 +10,26 @@ const triggerCourseStudentJoin = async (req, res) => {
         const { course, student } = req.body;
         //check if course and student exist
         const courseData = await Course.findById(course);
-        if (courseData.current_joined >= courseData.capacity) {
-            return res.status(400).json({
-                message: 'Course is full'
-            });
+        if (!courseData) {
+            if (courseData.current_joined >= courseData.capacity) {
+                return res.status(400).json({
+                    message: 'Course is full'
+                });
+            }
         }
 
         const studentData = await Student.findById(student);
         if (!courseData || !studentData) {
             return res.status(404).json({
                 message: 'Course or student not found'
+            });
+        }
+
+        //check if student already joined the course
+        const courseStudentData = await CourseStudent.findOne({ student: student, course: course });
+        if (courseStudentData) {
+            return res.status(400).json({
+                message: 'Student already joined the course'
             });
         }
 
@@ -57,6 +67,14 @@ const triggerCourseStudentLeave = async (req, res) => {
             });
         }
 
+        //check if student already joined the course
+        const courseStudentData = await CourseStudent.findOne({ student: student, course: course });
+        if (!courseStudentData) {
+            return res.status(400).json({
+                message: 'Student not joined the course'
+            });
+        }
+
         //if yes, delete the courseStudent
         const courseStudent = await CourseStudent.findOneAndDelete({
             student: student,
@@ -84,7 +102,7 @@ const triggerCourseStudentLeave = async (req, res) => {
 
 const attendanceHandler = async (req, res) => {
     try {
-        const { course, students } = req.body;
+        const { course, students, day } = req.body;
 
         // Check if course exists
         const courseData = await Course.findById(course);
@@ -95,7 +113,7 @@ const attendanceHandler = async (req, res) => {
         }
 
         // Extract student IDs
-        const studentIds = students.map(student => student._id);
+        const studentIds = students.map(student => student.id);
 
         // Check if all students exist
         const studentData = await Student.find({ _id: { $in: studentIds } });
@@ -108,12 +126,12 @@ const attendanceHandler = async (req, res) => {
         // Create bulk operations for updating attendance
         const bulkOperations = students.map(student => ({
             updateOne: {
-                filter: { student: student._id, course: course },
-                update: { 
-                    $inc: { 
-                        attendance_count: student.is_attended ? 1 : 0, 
-                        absent_count: student.is_attended ? 0 : 1 
-                    } 
+                filter: { student: student.id, course: course },
+                update: {
+                    $inc: {
+                        attendance_count: student.is_attended ? 1 : 0,
+                        absent_count: student.is_attended ? 0 : 1
+                    }
                 }
             }
         }));
@@ -129,15 +147,43 @@ const attendanceHandler = async (req, res) => {
         }
 
         //update attendance of each student
-        for(let student of students){
+        for (let student of students) {
             await Attendance.create({
-                student: student._id,
+                student: student.id,
                 course: course,
-                day: student.day,
-                status: student.is_attended ? 'no reason' : student.reasons
-            })
-        }
+                isAttend: student.is_attended,
+                day: day,
+                status: student.is_attended ? 'well done' : student.reasons
+            });
+        };
 
+        /* //send zalo message to parent if student is absent
+        for(let student of students){
+            if(!student.is_attended){
+                const studentData = await Student.findById(student.id).populate({
+                    path: 'parent',
+                    populate: {
+                        path: 'account',
+                        select: 'facebook',
+                        model: 'Account'
+                    }
+                });
+                const facebook = studentData.parent.account.facebook;
+                if(!facebook) {
+                    const facebookId = facebook.split('/')[3];
+                    console.log(facebookId);
+                    //send message to parent
+                    try{
+                        const sendMessageResponse = await sendFacebookMessage('Your child is absent today', phone);
+                        console.log('Facebook message sent:', sendMessageResponse);
+                    } catch (err) {
+                        console.error('Error sending zalo message:', err);
+                    }
+                } else {
+                    console.log('Parent does not have facebook account');
+                }
+            }
+        } */
         return res.status(200).json({
             data: bulkWriteResult,
             message: 'Attendance updated successfully'
@@ -153,7 +199,7 @@ const attendanceHandler = async (req, res) => {
 
 //get the attendance of a student in a course
 const getAttendance = async (req, res) => {
-    try { 
+    try {
         const { student, course } = req.params;
         //join course_student to get all joined course id
         const courseStudent = await CourseStudent.find({ student: student, course: course });
@@ -167,6 +213,8 @@ const getAttendance = async (req, res) => {
         });
     }
 }
+
+
 
 
 module.exports = {
